update types in todo service

+++ todoservice/types/todo.go
@@ -1,18 +1,18 @@
 package types
 
 type Todo struct {
-	ID          int
+	ID          int64
 	Name        string
 	Description string
-	UserId      int
+	UserId      int64
 }
 
 type Item struct {
-	ID       int
+	ID       int64
 	Title    string
 	Priority int
-	UserId   int
-	TodoId   int
+	UserId   int64
+	TodoId   int64
 }
 
 type TodoWithItems struct {
